src/jobs/evaluation/deepeval_evaluator/requirements.txt
deepeval
litellm
boto3
pandas
python-dotenv
